+++ KalikoCMS.Admin/Admin/Content/PropertyType/UniversalDateTimePropertyEditor.ascx.designer.cs
@@ -10,7 +10,7 @@
 namespace KalikoCMS.Admin.Content.PropertyType {
     
     
-    public partial class DateTimePropertyEditor {
+    public partial class UniversalDateTimePropertyEditor {
         
         /// <summary>
         /// LabelText control.
@@ -47,5 +47,14 @@ namespace KalikoCMS.Admin.Content.PropertyType {
         /// To modify move field declaration from designer file to code-behind file.
         /// </remarks>
         protected global::System.Web.UI.WebControls.Label ErrorText;
+        
+        /// <summary>
+        /// UniversalDateField control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.HiddenField UniversalDateField;
     }
 }