+++ KalikoCMS.Engine/Core/EditablePage.cs
@@ -20,6 +20,7 @@
 namespace KalikoCMS.Core {
     using System;
     using System.Linq;
+    using System.Web;
     using AutoMapper;
     using Configuration;
     using Data;
@@ -129,14 +130,15 @@ namespace KalikoCMS.Core {
                     pageInstance = new PageInstanceEntity {
                         PageId = PageId,
                         LanguageId = LanguageId,
-                        CreatedDate = DateTime.Now
+                        CreatedDate = DateTime.Now.ToUniversalTime()
                     };
 
                     context.Add(pageInstance);
                 }
 
                 pageInstance.PageName = PageName;
-                pageInstance.UpdateDate = DateTime.Now;
+                pageInstance.UpdateDate = DateTime.Now.ToUniversalTime();
+                pageInstance.Author = HttpContext.Current.User.Identity.Name;
                 pageInstance.StartPublish = StartPublish;
                 pageInstance.StopPublish = StopPublish;
                 pageInstance.VisibleInMenu = VisibleInMenu;